requirements_desktop.txt
flask
flask-wtf
reportlab
openai-whisper
pydub
pywebview
pyinstaller
